+++ Sources/StreamChat/WebSocketClient/Events/NotificationEvents.swift
@@ -416,3 +416,38 @@ class NotificationInviteRejectedEventDTO: EventDTO {
         )
     }
 }
+
+/// Triggered when a channel is deleted, this event is delivered to all channel members
+public struct NotificationChannelDeletedEvent: ChannelSpecificEvent {
+    /// The cid of the deleted channel
+    let cid: ChannelId
+
+    /// The channel that was deleted
+    let channel: ChatChannel
+
+    /// The event timestamp.
+    let createdAt: Date
+}
+
+class NotificationChannelDeletedEventDTO: EventDTO {
+    let cid: ChannelId
+    let channel: ChannelDetailPayload
+    let createdAt: Date
+    let payload: EventPayload
+
+    init(from response: EventPayload) throws {
+        cid = try response.value(at: \.cid)
+        channel = try response.value(at: \.channel)
+        createdAt = try response.value(at: \.createdAt)
+        payload = response
+    }
+    
+    func toDomainEvent(session: DatabaseSession) -> Event? {
+        guard let channelDTO = session.channel(cid: channel.cid) else { return nil }
+        return NotificationChannelDeletedEvent(
+            cid: cid,
+            channel: channelDTO.asModel(),
+            createdAt: createdAt
+        )
+    }
+}